# src/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "BadBeatMods API",
    "description": "This isn't really fully complete, but its better than absolutely nothing.\n\nThis API documentation is automatically generated and therefor may not be 100% accurate and may be missing a few fields.",
    "version": "0.0.1"
  },
  "host": "bbm.saera.gay",
  "basePath": "/",
...
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [
    "application/json",
    "multipart/form-data",
    "application/x-www-form-urlencoded"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/auth": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get logged in user information.",
        "description": "Get user information.",
        "responses": {
          "200": {
            "description": "Returns user information."
          },
          "401": {
            "description": "Unauthorized."
          }
        }
      }
    },
    "/api/auth/logout": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout.",
        "description": "Logout.",
        "responses": {
          "200": {
            "description": "Logout successful."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
    "/api/auth/github/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/link/discord/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/admin/linkversions": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Mark all versions as compatible with another gameversion.",
        "description": "Link two versions together.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "The versions to link.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "version1": {
                  "type": "number",
                  "example": 1
                },
                "version2": {
                  "type": "number",
                  "example": 2
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  }
}